% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcast-package.R
\docType{package}
\name{outcast-package}
\alias{outcast}
\alias{outcast-package}
\title{outcast: OUTlier Classification And Sequential Trimming}
\description{
Exported Functions:
\itemize{
\item \code{find_elbow} - .
\item \code{find_gross} - .
\item \code{fit_gmm} - .
\item \code{outcast_gmm} - .
\item \code{outback_gmm} - .
\item \code{plot_gross} - .
\item \code{simulate_gmm} - .
}
}
\author{
\strong{Maintainer}: Ultán Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

}
\keyword{internal}
